+++ application/GlkStyle.h
@@ -11,16 +11,16 @@
 @interface GlkStyle : NSObject
 {
     NSMutableDictionary *dict;
-    int stylenumber;
-    int windowtype;
-    int enabled[stylehint_NUMHINTS];
-    int value[stylehint_NUMHINTS];
+    NSInteger stylenumber;
+    NSInteger windowtype;
+    NSInteger enabled[stylehint_NUMHINTS];
+    NSInteger value[stylehint_NUMHINTS];
 }
 
-- initWithStyle: (int)stylenumber_
-     windowType: (int)windowtype_
-	 enable: (int*)enablearray
-	  value: (int*)valuearray;
+- initWithStyle: (NSInteger)stylenumber_
+     windowType: (NSInteger)windowtype_
+	 enable: (NSInteger *)enablearray
+	  value: (NSInteger *)valuearray;
 - (void) prefsDidChange;
 - (NSDictionary*) attributes;
